+++ functional-tests/functional/input/src/cpp/include/ExternalTypes.h
@@ -20,6 +20,7 @@
 
 #pragma once
 
+#include <chrono>
 #include <string>
 #include <system_error>
 #include <vector>